+++ math_calc.h
@@ -12,6 +12,6 @@ public:
 	static sf::Vector2f Get_xy_SpeedRatio(sf::Vector2f _pos, sf::Vector2f _dirPos);	
 	static float MakePositive(float _num);
 	
-	static float GetAngle(sf::Vector2f _pos, sf::Vector2i _dirPos);
+	static float GetAngle(sf::Vector2f _pos, sf::Vector2f _dirPos);
 	
 };
